Implement Overdrive data objects using data classes
A data class will make a new instance out of an OD object and endow it
with the methods to normalize the object so that it will be easier to use.
Currently, these methods are defined as ad-hoc functions within the
od_api module, making it lengthy and hard to understand.

Signed-off-by: Steven Chan <steven3416@gmail.com>