Fix progress message in action dialogues
The jQuery ajax method does not implement the full promise interface; it
does not call notify and so the progress callback cannot work. As a
substitute, we pass the progress callback as the beforeSend callback,
which is executed before an ajax send.

Also, we use a progress bar instead of a progress text message.

Signed-off-by: Steven Chan <steven3416@gmail.com>