Replace progress text with a progress bar
As a way to do it, we slightly generalized the set_message method of the
action dialogue, to default to using a progress bar for the body text if
there is no text string specified as an argument nor as a body property
of the scenario. Thus, for an intent scenario, it is sufficient to use
the default body.

Signed-off-by: Steven Chan <steven3416@gmail.com>