Change how download button works to allow for right-clicking to work
On user clicking a download button, we convert the download URL to a content URL,
and then convert the button to a link with a new message. This will
allow the link to be used directly, ie, right-clicking to open in a new
tab.

The interaction is not perfect, for example, the content link will be
useful for only 60 seconds. A more elaborate interaction will need to be
devised.

Signed-off-by: Steven Chan <steven3416@gmail.com>