Refine the rendering of myopac account summary table and dashboard
Before, the table and dashboard were built to show empty e-item entries
that would sit there while the user waits for the several API requests
to finish before the total values were available to be placed in the
entries.  Now, the entries are built and shown after values are
available.

A solid gray bar shows in the dashboard during the wait interval. We
would have shown a progress bar, but it's too hard to get it to show
because of interference from the existing style sheet.

The first column of the summary table shows totals for physical items
and was suppressed until the arrival of totals for e-items, but in fact
the totals of physical items are already available from the old table.
Now, we show them immediately, and we don't show the second column at
all until it is filled with values.

Signed-off-by: Steven Chan <steven3416@gmail.com>